\name{twin-package}
\alias{twin-package}
\docType{package}
\title{
\packageTitle{twin}
}
\description{
\packageDescription{twin}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{twin}
\packageIndices{twin}
}
\author{
\packageAuthor{twin}

Maintainer: \packageMaintainer{twin}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
