\name{twin-package}
\alias{twin-package}
\docType{package}
\title{
\packageTitle{twin}
}
\description{
\packageDescription{twin}
}
\details{
    Estimating parameters of the Time-Window of Integration (TWIN)
    model and simulating data for parameter sets under the model. Multisensory
    integration of sensory information into a perceptual unit (event) occurs only
    if the peripheral sensory processes all terminate within a given temporal
    interval, the Time-Window of Integration (TWIN). Assumed are two stages. In the
    first stage, there is a (stochastically independent) race among the peripheral
    processes. The second stage comprises all processes following the first stage
    including preparation and execution of a response. The parameters are the
    processing times. Other parameters are the length of the TWIN and the amount of
    integration. Data are reaction times.
}
\author{
\packageAuthor{twin}

Maintainer: \packageMaintainer{twin}
}
\references{
    Colonius, H., & Diederich, A. (2004). Multisensory interaction in saccadic
    reaction time: a time-window-of-integration model. \emph{Journal of
        cognitive neuroscience, 16}, \bold{6}, 1000--1009.
    \doi{10.1162/0898929041502733}

    Diederich, A., & Colonius, H. (2015). The time window of multisensory
    integration: Relating reaction times and judgments of temporal order.
    \emph{Psychological review, 122}, \bold{2}, 232.
    \doi{10.1037/a0038696}

    Kandil, F. I., Diederich, A., & Colonius, H. (2014). Parameter recovery for
    the time-window-of-integration (twin) model of multisensory integration in
    focused attention. \emph{Journal of vision, 14}, \bold{11}, 14-14.
    \doi{10.1167/14.11.14}
}
\keyword{ package }
